refactor: remove networkType prop from APIs
