+++ packages/btc/src/address.ts
@@ -1,8 +1,17 @@
 import { bitcoin } from './bitcoin';
-import { AddressType } from './types';
 import { NetworkType, toPsbtNetwork } from './network';
 import { ErrorCodes, TxBuildError } from './error';
 import { remove0x, toXOnly } from './utils';
+
+export enum AddressType {
+  P2PKH,
+  P2WPKH,
+  P2TR,
+  P2SH_P2WPKH,
+  P2WSH,
+  P2SH,
+  UNKNOWN,
+}
 
 /**
  * Check weather the address is supported as a from address.
