+++ packages/btc/src/constants.ts
@@ -12,6 +12,6 @@ export const BTC_UTXO_DUST_LIMIT = 1000;
 export const RGBPP_UTXO_DUST_LIMIT = 546;
 
 /**
- * An empty placeholder, filled with 0s for the txid of the BTC transaction.
+ * An empty 32-byte placeholder, filled with 0s for the txid of the BTC transaction.
  */
 export const BTC_TX_ID_PLACEHOLDER = '0x' + '0'.repeat(64);
