+++ packages/btc/src/types.ts
@@ -1,3 +1,5 @@
+import { AddressType } from './address';
+
 export interface Utxo {
   txid: string;
   vout: number;
@@ -6,14 +8,4 @@ export interface Utxo {
   addressType: AddressType;
   scriptPk: string;
   pubkey?: string;
-}
-
-export enum AddressType {
-  P2PKH,
-  P2WPKH,
-  P2TR,
-  P2SH_P2WPKH,
-  P2WSH,
-  P2SH,
-  UNKNOWN,
 }
